+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [3.7.3]
+
+### Changed
+
+- [TD-1721] template is not updated if updated_at does not change 
+
 ## [3.7.2]
 
 ### Added